+++ src/main/java/com/g6jamm/stima/data/repository/mysql/util/type/ConnectionType.java
@@ -1,0 +1,6 @@
+package com.g6jamm.stima.data.repository.mysql.util.type;
+
+public enum ConnectionType {
+  HEROKU,
+  LOCAL
+}
