--- src/main/java/com/g6jamm/stima/data/repository/util/type/ConnectionType.java
@@ -1,6 +1,0 @@
-package com.g6jamm.stima.data.repository.util.type;
-
-public enum ConnectionType {
-  HEROKU,
-  LOCAL
-}
